Test restructure step 5c: Switch/Elements/Configure proof conversions
Applies the DaemonToolCalling conversion pattern from step 5b (Touch)
to three more commands sharing the identical shape and fake surface:
SwitchCommand, ElementsCommand, ConfigureCommand. Each command's run()
had its withDaemonClient closure body extracted into execute(on: any
DaemonToolCalling), with validation logic staying in run() before the
daemon call. Each gets one new logic-test file converting its "no
session" test to run in-process against the existing FakeDaemonClient
instead of spawning a subprocess + daemon. Real subprocess coverage
that depends on genuine daemon-side behavior (happy-path round trips,
out-of-range errors, macOS-session policy rejection) is untouched.

Also fixes SwitchCommand.execute(on:)'s doc comment, which was missing
the "callers must have already validated" precondition trailer that
Touch's and Configure's already carry.
